% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyCars.R
\docType{data}
\name{emptyCars}
\alias{emptyCars}
\title{NA values in mtcars dataset}
\format{
A data frame with 32 rows and 11 variables:
\describe{
\item{\code{mpg}}{double. miles per gallon.}
\item{\code{cyl}}{double. cylinder of car.}
\item{\code{disp}}{double. displacement.}
\item{\code{hp}}{double. gross horsepower.}
\item{\code{drat}}{double. rear axle ratio.}
\item{\code{wt}}{double. weight over 1000.}
\item{\code{qsec}}{double. 1/4 mile time.}
\item{\code{vs}}{double. variable signaling.}
\item{\code{am}}{double. transmission type.}
\item{\code{gear}}{double. number of forward gears.}
\item{\code{carb}}{double. number of carburetors.}
}
}
\usage{
emptyCars
}
\description{
Using the mtcars dataset, and plugging in 100 NA values randomly throughout the data set.
}
\keyword{datasets}
