% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixNA.R
\name{fixNA}
\alias{fixNA}
\title{fixNA}
\usage{
fixNA(data = mtcars, value_NA = c(99, NA), recode_NA = NA)
}
\arguments{
\item{data}{an object of class data.frame to transform.}

\item{value_NA}{single element or vector of elements to transform from. Default is 99 and NA.}

\item{recode_NA}{value to transform to. Default is NA.}
}
\value{
an object of class data.frame
}
\description{
\code{fixNA} takes a data frame and transforms value_NA into recode_NA.
}
\examples{
# data("randomCars")
# fixNA(randomCars)
}
