% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{ichor_provenance}
\alias{ichor_provenance}
\title{Summarize input provenance}
\usage{
ichor_provenance(x)
}
\arguments{
\item{x}{An \code{ichor_sample} or \code{ichor_cohort}.}
}
\value{
A data frame of source files, sizes, and MD5 checksums.
}
\description{
Summarize input provenance
}
