% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ichorViz-package.R
\docType{package}
\name{ichorViz-package}
\alias{ichorViz}
\alias{ichorViz-package}
\title{ichorViz: explore and compare ichorCNA profiles}
\description{
ichorViz consumes output generated by ichorCNA. It provides validated data
objects, genome-wide and regional plots, paired or multi-sample comparisons,
and cohort-scale matrices and heatmaps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bauernhofere/ichorViz}
  \item Report bugs at \url{https://github.com/bauernhofere/ichorViz/issues}
}

}
\author{
\strong{Maintainer}: Emil Bauernhofer \email{emil.bauernhofer@gmail.com}

}
\keyword{internal}
