% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeNA-package.R
\docType{package}
\name{seeNA-package}
\alias{seeNA}
\alias{seeNA-package}
\title{seeNA: explore and compare ichorCNA profiles}
\description{
seeNA consumes output generated by ichorCNA. It provides validated data
objects, genome-wide and regional plots, paired or multi-sample comparisons,
and cohort-scale matrices and heatmaps.
}
\section{Scientific methods}{

The installed \code{methods.md} documents coordinate conventions, supported
schemas, categorical aggregation, coverage, privacy and limitations:
\code{file.show(system.file("methods.md", package = "seeNA"))}.
}

\examples{
root <- system.file("extdata", package = "seeNA")
s <- read_ichor_sample(file.path(root, "example-a.cna.seg"),
                       genome_build = "hg38")
plot_ichor_profile(s)
cohort <- read_ichor_cohort(file.path(root, "example-manifest.csv"), "hg38")
calls <- ichor_matrix(cohort, value = "call", chromosomes = "1")
calls$values[, 1:4]

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bauernhofere/seeNA}
  \item Report bugs at \url{https://github.com/bauernhofere/seeNA/issues}
}

}
\author{
\strong{Maintainer}: Emil Bauernhofer \email{emil.bauernhofer@gmail.com}

}
\keyword{internal}
