% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeNA.R
\name{seeNA}
\alias{seeNA}
\title{seeNA}
\usage{
seeNA(data, color = "magenta")
}
\arguments{
\item{data}{an object of class data.frame to visualize.}

\item{color}{color of points for visualization. Default is the color magenta.}
}
\description{
\code{seeNA} takes a data frame and displays a visualization of the distribution of missing values within your data frame.
}
\examples{
# data(mtcars)
# x = makeNA(mtcars, n_missing = 50)
# seeNA(x)
}
