% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryNA.R
\name{summaryNA}
\alias{summaryNA}
\title{summaryNA}
\usage{
summaryNA(df, round = 0)
}
\arguments{
\item{df}{an object of class data.frame.}

\item{round}{rounding for summary statistics}
}
\value{
an object of class matrix
}
\description{
\code{summaryNA} takes a data frame and prints out summary statistics related to the amount of missing values in a data frame.
}
\examples{
# library(seeNA)
# data("emptyCars")
# summaryNA(emptyCars)
}
