% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeNA.R
\name{imputeNA}
\alias{imputeNA}
\title{imputeNA}
\usage{
imputeNA(data = emptyCars, function.imp = "mean", ...)
}
\arguments{
\item{data}{an object of class data.frame to impute missing values.}

\item{...}{additional arguments passed to the function.}

\item{package}{An optional character string naming the imputation function. By default, the imputation function is mean.}
}
\value{
list or data.frame object
}
\description{
\code{imputeNA} imputes the missing values in a dataframe based on the specified imputation function.
}
\examples{
# library(seeNA)
# data("emptyCars")
# imputeNA(emptyCars)
}
