% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ichor_profile}
\alias{plot_ichor_profile}
\title{Plot one genome-wide ichorCNA profile}
\usage{
plot_ichor_profile(
  x,
  call_column = "corrected_call",
  colors = ichor_state_colors(),
  point_size = 0.35
)
}
\arguments{
\item{x}{An \code{ichor_sample}.}

\item{call_column}{Column used to color copy-number states.}

\item{colors}{Named state color vector.}

\item{point_size}{Bin point size.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot one genome-wide ichorCNA profile
}
