% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ichor_region}
\alias{plot_ichor_region}
\title{Plot a genomic region}
\usage{
plot_ichor_region(samples, region, colors = NULL, point_size = 0.55)
}
\arguments{
\item{samples}{An \code{ichor_sample} or list of samples.}

\item{region}{Required chromosome or \code{chr:start-end} interval.}

\item{colors}{Optional named vector of sample colors.}

\item{point_size}{Bin point size.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a genomic region
}
