% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{read_ichor_sample}
\alias{read_ichor_sample}
\title{Construct a validated ichorCNA sample}
\usage{
read_ichor_sample(
  cna_seg,
  seg = NULL,
  params = NULL,
  genome_build,
  sample_id = NULL
)
}
\arguments{
\item{cna_seg}{Path to a \code{.cna.seg} file.}

\item{seg}{Optional path to a \code{.seg} file.}

\item{params}{Optional path to a \code{.params.txt} file.}

\item{genome_build}{Genome build, currently \code{"hg19"} or \code{"hg38"}.}

\item{sample_id}{Optional sample identifier.}
}
\value{
An object of class \code{ichor_sample}.
}
\description{
Construct a validated ichorCNA sample
}
