% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_ichor_segments}
\alias{read_ichor_segments}
\title{Read an ichorCNA segment file}
\usage{
read_ichor_segments(path)
}
\arguments{
\item{path}{Path to an ichorCNA \code{.seg} file.}
}
\value{
A data frame with normalized segment columns.
}
\description{
Read an ichorCNA segment file
}
