% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{validate_ichor_cohort}
\alias{validate_ichor_cohort}
\title{Validate a cohort}
\usage{
validate_ichor_cohort(x)
}
\arguments{
\item{x}{An \code{ichor_cohort}.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate a cohort
}
