% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{validate_ichor_matrix}
\alias{validate_ichor_matrix}
\title{Validate a cohort matrix}
\usage{
validate_ichor_matrix(x)
}
\arguments{
\item{x}{An \code{ichor_matrix}.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate a cohort matrix
}
