% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{validate_ichor_sample}
\alias{validate_ichor_sample}
\title{Validate an ichorCNA sample}
\usage{
validate_ichor_sample(x)
}
\arguments{
\item{x}{An \code{ichor_sample} object.}
}
\value{
\code{x}, invisibly. Invalid samples raise an error.
}
\description{
Validate an ichorCNA sample
}
